Phase_1_Speech_Analysis/requirements.txt
openai-whisper
librosa
moviepy
numpy
